Adhere to contributing.md
