+++ Validator/Validator.csproj
@@ -44,6 +44,7 @@
     <Compile Include="IpVersion.cs" />
     <Compile Include="IsBase64.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ValidationExtensions.cs" />
     <Compile Include="Validator.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
